Initial commit
Initial commit of Cucumber X Selenium project using a mvn boilerplate. Added Page structure and feature file strucutre. Completed A simple search feature to ensure all the correct dependencies are working together. For ease of running i chose maven so 'mvn test' can be run by the assessor. As a result of building in mvn i've faced a dependency management nightmare and opted to use a library that managed webdriver versions to match the selenium-java dependency